stock-alerts/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
httpx==0.27.2
apscheduler==3.10.4
python-dotenv==1.0.1
